Making it more organised
