+++ BasePage.cs
@@ -1,4 +1,5 @@
 using System.Windows.Controls;
+using Win_Labs.Themes;
 
 namespace Win_Labs
 {
