+++ Settings/AppSettingsManager.cs
